Restructures directory
