+++ client/src/components/Businesses/smart/UpdateBusiness.js
@@ -3,7 +3,7 @@ import { Input } from 'react-materialize';
 import { Link } from 'react-router-dom';
 import { NotificationManager, NotificationContainer } from 'react-notifications';
 import alertify from 'alertifyjs';
-import Errors from '../Messages/Errors';
+import Errors from '../../Messages/presentational/Errors';
 
 /**
  *
@@ -72,10 +72,9 @@ export default class UpdateBusiness extends React.Component {
       }
       this.props.updateBusiness(this.props.match.params.id, businessDetails)
         .then((response) => {
-        //   NotificationManager.success('Business Updated', 'Successful');
           alertify.set('notifier', 'position', 'top-right');
           alertify.success('Business Updated Successfully');
-          setTimeout(() => this.props.history.push(`/businessProfile${this.props.match.params.id}`), 4000);
+          setTimeout(() => this.props.history.push(`/businessProfile${this.props.match.params.id}`), 2000);
         })
         .catch((error) => {
           window.scroll(0, 0);
