+++ client/src/components/Home/presentational/HomeImageSlider.js
