+++ client/src/components/Home/presentational/HowItWorks.js
