+++ client/src/components/NavBar/presentational/GuestNavBar.js
